Add timestamp of next non-boot event to DBUS signal next_bootup_event
The signal has two parameters, int next_boot_event, and
int next_non_boot_event. Both are timestamps in the form of
seconds since the Unix epoch. The former gives the trigger
date of the next event with the boot flag enabled, and the
latter the trigger date of the next event without the boot
flag enabled.

While the device is up and running, DSME uses both parameters
to program the RTC to make the device wake up from suspend
to handle events.

At shutdown DSME programs the RTC to power up the device to
trigger the next event with the boot flag set, whose trigger
date is defined by the paramter next_boot_event.