[timed] Change location of time zone link that timed manages
Timed links time zone files in /usr/share/zoneinfo to
/var/lib/timed, for example

  /var/lib/timed/localtime -> /usr/share/zoneinfo/Europe/Helsinki

The system time zone /etc/localtime is in turn linked to
/var/lib/timed/localtime, so that it follows the time zone
set by timed.

The directory /var/lib/timed can be modified by root and
members of group timed, timed installation creates group
timed and adds the default user to the group.

These changes are required as timed is not run with root
privileges, and it does not have the rights to change
/etc/localtime.